Refactored login page
