+++ lib/ui/page/login/login_page.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/material.dart';
+import 'package:labs/ui/page/login/login_service.dart';
+
+class LoginPage extends StatefulWidget {
+  const LoginPage({super.key});
+
+  @override
+  LoginPageState createState() => LoginPageState();
+}
+
+class LoginPageState extends State<LoginPage> {
+  final _emailController = TextEditingController();
+  final _passwordController = TextEditingController();
+  late LoginService _loginService;
+
+  @override
+  void initState() {
+    super.initState();
+    _loginService = LoginService(context);
+    _loginService.initializeApp();
+  }
+
+  @override
+  void dispose() {
+    _loginService.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Login'),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            TextField(
+              controller: _emailController,
+              decoration: const InputDecoration(labelText: 'Email'),
+            ),
+            TextField(
+              controller: _passwordController,
+              obscureText: true,
+              decoration: const InputDecoration(labelText: 'Password'),
+            ),
+            const SizedBox(height: 16),
+            ElevatedButton(
+              onPressed: () => _loginService
+                  .loginUser(_emailController.text, _passwordController.text),
+              child: const Text('Login'),
+            ),
+            TextButton(
+              onPressed: () => Navigator.pushNamed(context, '/register'),
+              child: const Text('Don\'t have an account? Register'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
